% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{death}
\alias{death}
\title{Confirmed COVID-19 deaths}
\format{
a list with
\itemize{
\item{Province_State}{ name of the US state}
\item{date ...}{ a column for each date}
}
}
\description{
Confirmed COVID-19 deaths in US states
}
\keyword{data}
