% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pi.SIR.R
\name{pi.SIR}
\alias{pi.SIR}
\title{SIR model with a time-varying transmission rate}
\usage{
pi.SIR(Y, R, pi_qbar0 = NULL, change_time = NULL,
  exponential = FALSE, lambda0 = NULL, begin_str = "01/23/2020",
  T_fin = 200, nchain = 4, nadapt = 10000, M = 5000, thn = 10,
  nburnin = 2000, dic = FALSE, file_add = character(0),
  save_files = FALSE, death_in_R = 0.02, casename = "pi.SIR",
  beta0 = 0.2586, gamma0 = 0.0821, R0 = beta0/gamma0,
  gamma0_sd = 0.1, R0_sd = 1)
}
\arguments{
\item{Y}{the observed infected proportions by time.}

\item{R}{the observed removed proportions by time, including death and recovered.}

\item{pi_qbar0}{the time dependent function \eqn{\pi_bar{q}(t)} between 0 and 1.}

\item{change_time}{the change time points for step function pi, defalt value is \code{NULL}.}

\item{exponential}{logical, whether \eqn{\pi_bar{q}(t)} is exponential \eqn{\exp(-\lambda_0t)} or not; the default is \code{FALSE}.}

\item{lambda0}{the rate of decline in the exponential function in \eqn{\exp(-\lambda_0t)}.}

\item{begin_str}{the character of the starting time, the default is "01/23/2020", which is the starting date that the local government blocked Wuhan City.}

\item{T_fin}{the maximum follow-up date after the beginning \code{begin_str}, the default is 200.}

\item{nchain}{the number of MCMC chains called in rjags, the default is 4.}

\item{nadapt}{the number of iterations for adaptation in the MCMC. The default is 1e4, which is what we suggest to use 1e4 instead.}

\item{M}{number of draws from each chain, without considering thinning. The default is M=5e3 but we suggest using 5e5 instead.}

\item{thn}{thinning interval for monitors. Thus, the total number of draws would be \code{round(M/thn)*nchain }. The default is 10.}

\item{nburnin}{the burn-in period. The default is nburnin=2e3 but we suggest using nburin=2e5.}

\item{dic}{logical, whether compute the DIC or deviance information criterion.}

\item{file_add}{the string to denote the location to save the output files and tables.}

\item{save_files}{logical, whether save (\code{TRUE}) the results or not (\code{FALSE}). This will enable saving the summary table, trace plots, and the plot of the posterior mean of the first derivative of the infection proportion \eqn{\theta_t^I}.}

\item{death_in_R}{numeric value of average ratio between deaths and cumulative removed subjects. The default is 0.4 within Hubei, and 0.02 outsite Hubei according to the reported data by Feb 11, 2020.}

\item{casename}{string of the job's name. The default is "pi.SIR".}

\item{beta0}{the hyperparameter of the mean transmission rate, the default is the one estimated from SARS (0.2586) first-month outbreak.}

\item{gamma0}{the hyperparameter of the mean recovery rate (including death), the default is estimated from SARS (0.0821) first-month outbreak.}

\item{R0}{the hyperparameter of the mean R0 value. The default is \code{beta0/gamma0}, which can be overwritten by discarding the value set in \code{beta0}.}

\item{gamma0_sd}{the standard deviance for the prior of the recovery/remove rate, the default is 0.1.}

\item{R0_sd}{the standard deviance for the prior of R0, the default is 1.}
}
\value{
\item{casename}{casename defined before}
\item{incidence_mean}{mean incidence}
\item{incidence_ci}{2.5\%, 50\%, and 97.5\% quantiles of the incidences}
\item{out_table}{summary table with varibles including the posterior mean of the proportions of the 3 states at their last observation date, and their respective credible inctervals (ci) including the median; the mean and ci of the reporduction number (R0), removed/recovery rate (gamma), transmission rate  (beta)}
\item{forecast_infection}{plot to forecast the infection with following lines: the vertial blue line denotes the last observation date; vertial purple line denotes the change point indicating a decrease in infection proportion or the date with 0 value of the posterior mean first-derivative infection proportion \eqn{\theta_t^{\prime I}}; the vertial darkgreen line denotes the deacceleration point of the increasing infection proportion or the date with maximum value of the posterior mean first-derivative infection proportion \eqn{\theta_t^{\prime I}}; the darkgray line denotes the posterior mean of the infection proportion; the red line denotes the posterior median of the infection proportion}
\item{forecast_removed}{plot to forecast the removed lines described in \code{forecast_infection}. The meaning of the vertical lines were identical, but the horizontal mean and median were corresponding to the posterior mean and median of the removed state. Moreover, we introduce an additional line for the estimated death proportion, which is based on the input \code{death_in_R}}
\item{first_stat_mean}{the mean first stationary date, which is the change point that we observe decline in the infection proportion (\eqn{\theta_t^I}), or its stationary point; it is calculated using the average of the 0-points of all the repeats of posterior draws of the first derivative proportion or \eqn{\theta_t{\prime I}}; this value may be slightly different from the one labeled by the "purple" lines in the \code{forecast_infection} and \code{forecast_removed} two plots, as the latter indicate the 0-value point of the first-derivative the posterior mean of \eqn{\theta_t^I}.}
\item{first_stat_ci}{following the definition of \code{first_stat_mean}, but is the corresponding credible interval.}
\item{second_stat_mean}{the mean second stationary date, which is the change point that we observe the decline in the increasing spead of the infection proportion (\eqn{\theta_t^I}) or its derivative's stationary point;it is calculated using the average of the stationary values of all the repeats of posterior draws of the first-derivative porportion of infection or \eqn{\theta_t^{\prime I}}; this value may be slightly different from the one labeled by the "darkgreen" lines in the \code{forecast_infection} and \code{forecast_removed} two plots, as the latter indicate the stationary point of the first-derivative the posterior mean of \eqn{\theta_t^I}.}
\item{dic_val}{the output of \code{dic.sample()} in \code{rjags}, computing deviance information criterion for model comparison.}
}
\description{
SIR model with fixed and known changes in the transmission rate, either stepwise or continuous.
}
\details{
In this function we introduce a time-dependent multiplier (boundared between 0 and 1) of the transmisison rate, \eqn{pi_{qbar}(t)}. In this way,  we can edow the transmission some time-dependent changes, either following a step-function or a smooth exponetial function (\eqn{\exp(\lambda_0t)}). The parameters of the function and change points, if any, need to be predefined.
}
\examples{
NI_complete <- c( 41,41,41,45,62,131,200,270,375,444,549, 729,1052,1423,2714,3554,4903,5806,7153,9074,11177,13522,16678,19665,22112,24953,27100,29631,31728,33366)
RI_complete <- c(1,1,7,10,14,20,25,31,34,45,55,71,94,121,152,213,252,345,417,561,650,811,1017,1261,1485,1917,2260,2725,3284,3754)
N=58.5e6
R <- RI_complete/N
Y <- NI_complete/N- R #Jan13->Feb 11
### Step function of pi_qbar(t)
change_time <- c("01/23/2020","02/04/2020","02/08/2020")
pi_qbar0 <- c(1.0,0.9,0.5,0.1)
res.step <-pi.SIR(Y,R,begin_str="01/13/2020",T_fin=200,pi_qbar0=pi_qbar0,change_time=change_time,casename="Hubei_step",save_files = T)
res.step$forecast_infection
### continuous exponential function of pi_qbar(t)
res.exp <- pi.SIR(Y,R,begin_str="01/13/2020",T_fin=200,pi_qbar0=pi_qbar0,change_time=change_time,exponential=TRUE,lambda0=0.01,casename="Hubei_exp")
res.exp$forecast_infection
### without pi_qbar(t)
res.nopi <- pi.SIR(Y,R,begin_str="01/13/2020",T_fin=200,casename="Hubei_nopi")
res.nopi$forecast_infection



}
