% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{USA_state_N}
\alias{USA_state_N}
\title{Population of US states}
\format{
a list with
\itemize{
\item{Province_State}{ State as a character string}
\item{N}{ Population of the state}
}
}
\description{
Data frame with the populations of each state.
}
\keyword{data}
